update to correctly match age with range in different combinations
characteristics[age] with range annotation matching is broken with the current regex. This update change the regex to be a bit more complex. Add PXD001474 annotated sdrf file from the bigbio/proteomics-sample-metadata as reference for testing with range.

Also update a docstring to correctly refer to the input parameter name in OntologyValidator.validate