adding validation proof utility function and testing
adding a ValidationProof class that can be used to generate a SHA512 hash for the validation action that is combination of content of the sdrf, content of the template, version of the  validator, timestamp and an optional unique salt string. validate_sdrf function from parse_sdrf also have optional parameters for producing a validation proof json and/or just print the hash into the cli.